refactor: replace extension traits by dedicated types
